requirements.txt
loguru==0.6.0
psutil==5.9.0
pyngrok==5.1.0
pypresence==4.2.1
PyYAML==6.0
